backend/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
pydantic==2.12.5
python-dotenv==1.2.2
openai==2.29.0
httpx==0.28.1
sqlalchemy==2.0.48
psycopg[binary]==3.3.3
